--- .gitignore
@@ -1,1 +1,0 @@
-.ipynb_checkpoints/
